# dist/config/swagger-output.json
{
    "swagger": "2.0",
    "info": {
        "title": "Jiffy E-Commerce API",
        "version": "1.0.0",
        "description": "Complete API documentation for Jiffy E-Commerce Platform - Buyer and Seller APIs"
    },
    "host": "localhost:3000",
    "basePath": "/",
...
    "schemes": [
        "http",
        "https"
    ],
    "securityDefinitions": {
        "bearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "scheme": "bearer",
            "bearerFormat": "JWT",
            "in": "header"
        }
    },
    "paths": {
        "/api/auth/send-otp": {
            "post": {
                "description": "",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "phone_number": {
                                    "example": "any"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
        "/api/auth/verify-otp": {
            "post": {
                "description": "",
                "parameters": [
                    {
                        "name": "user-agent",
                        "in": "header",
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "phone_number": {
                                    "example": "any"
                                },
                                "otp": {
                                    "example": "any"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
        "/api/auth/refresh-token": {
            "post": {
                "description": "",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "refreshToken": {
                                    "example": "any"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "description": "",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "refreshToken": {
                                    "example": "any"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
        "/api/auth/register-seller": {
            "post": {
                "description": "",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "phone_number": {
                                    "example": "any"
                                },
                                "email": {
                                    "example": "any"
                                },
                                "password": {
                                    "example": "any"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
        "/api/auth/verify-seller-otp": {
            "post": {
                "description": "",
                "parameters": [
                    {
                        "name": "user-agent",
                        "in": "header",
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "otp": {
                                    "example": "any"
                                },
                                "phone_number": {
                                    "example": "any"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
...
    }
}